requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
sqlalchemy==2.0.48
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.22
pydantic==2.12.5
email-validator==2.3.0
python-dotenv==1.1.1
pytest==8.4.2
httpx==0.28.1
alembic==1.14.1
